feat: Add a sidebar with larivaar assist switch
