+++ components/index.ts
@@ -1,0 +1,5 @@
+import Ang from './Ang';
+import Settings from './Settings/setting';
+import Launchpad from './Launchpad/lauchpad';
+
+export {Ang, Settings, Launchpad};
